+++ build-aux/vs2022/libbitcoin_server/libbitcoin_server.vcxproj
@@ -124,6 +124,7 @@
     <ClCompile Include="..\..\..\src\mnode\mnode-validation.cpp" />
     <ClCompile Include="..\..\..\src\mnode\rpc\mnode-rpc-utils.cpp" />
     <ClCompile Include="..\..\..\src\mnode\rpc\pastelid-rpc.cpp" />
+    <ClCompile Include="..\..\..\src\mnode\rpc\governance-rpc.cpp" />
     <ClCompile Include="..\..\..\src\mnode\rpc\storage-fee.cpp" />
     <ClCompile Include="..\..\..\src\mnode\rpc\tickets-activate.cpp" />
     <ClCompile Include="..\..\..\src\mnode\rpc\tickets-fake.cpp" />
@@ -213,6 +214,7 @@
     <ClInclude Include="..\..\..\src\mnode\rpc\masternodebroadcast.h" />
     <ClInclude Include="..\..\..\src\mnode\rpc\mnode-rpc-utils.h" />
     <ClInclude Include="..\..\..\src\mnode\rpc\pastelid-rpc.h" />
+    <ClInclude Include="..\..\..\src\mnode\rpc\governance-rpc.h" />
     <ClInclude Include="..\..\..\src\mnode\rpc\storage-fee.h" />
     <ClInclude Include="..\..\..\src\mnode\rpc\tickets-activate.h" />
     <ClInclude Include="..\..\..\src\mnode\rpc\tickets-fake.h" />
