+++ build-aux/vs2022/libbitcoin_server/libbitcoin_server.vcxproj.filters
@@ -301,6 +301,9 @@
     <ClCompile Include="..\..\..\src\mnode\rpc\tickets-findbylabel.cpp">
       <Filter>Source Files\mnode\rpc</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\..\src\mnode\rpc\governance-rpc.cpp">
+      <Filter>Source Files\mnode\rpc</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\..\src\addrman.h">
@@ -643,6 +646,9 @@
       <Filter>Source Files\mnode\tickets</Filter>
     </ClInclude>
     <ClInclude Include="..\..\..\src\mnode\rpc\tickets-findbylabel.h">
+      <Filter>Source Files\mnode\rpc</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\src\mnode\rpc\governance-rpc.h">
       <Filter>Source Files\mnode\rpc</Filter>
     </ClInclude>
   </ItemGroup>
