+++ qa/pull-tester/rpc-tests.sh
@@ -110,7 +110,6 @@ declare -a testScriptsMN=(
     'secure_container.py'
     'mn_tickets_username_change.py'
     'mn_tickets_validation.py'
-    'mn_governance.py'
     'mn_messaging.py'
     'mn_payment.py'
     'mn_bugs.py'
@@ -119,6 +118,10 @@ declare -a testScriptsMN=(
 declare -a testScriptsMNslow=(
     'mn_main.py'
     'mn_tickets.py'
+)
+
+declare -a testScriptsMNdisabled=(
+    'mn_governance.py'
 )
 
 declare -a testScriptsExt=(
