+++ src/mnode/mnode-notificationinterface.cpp
@@ -1,13 +1,14 @@
 // Copyright (c) 2014-2017 The Dash Core developers
-// Copyright (c) 2018 The Pastel developers
+// Copyright (c) 2018-2022 The Pastel developers
 // Distributed under the MIT software license, see the accompanying
-// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+// file COPYING or https://www.opensource.org/licenses/mit-license.php.
 
-#include "main.h"
-#include "chainparams.h"
+#include <main.h>
+#include <chainparams.h>
 
-#include "mnode/mnode-notificationinterface.h"
-#include "mnode/mnode-controller.h"
+#include <mnode/mnode-notificationinterface.h>
+#include <mnode/mnode-controller.h>
+#include <mnode/tickets/ticket-types.h>
 
 void CACNotificationInterface::InitializeCurrentBlockTip()
 {
@@ -36,5 +37,7 @@ void CACNotificationInterface::UpdatedBlockTip(const CBlockIndex *pindexNew, boo
 	
 	masterNodeCtrl.masternodeManager.UpdatedBlockTip(pindexNew);
 	masterNodeCtrl.masternodePayments.UpdatedBlockTip(pindexNew);
+#ifdef GOVERNANCE_TICKETS
 	masterNodeCtrl.masternodeGovernance.UpdatedBlockTip(pindexNew);
+#endif // GOVERNANCE_TICKETS
 }
