+++ src/mnode/mnode-notificationinterface.h
@@ -1,9 +1,9 @@
 #pragma once
 // Copyright (c) 2015 The Bitcoin Core developers
-// Copyright (c) 2018-2021 The Pastel Core developers
+// Copyright (c) 2018-2022 The Pastel Core developers
 // Distributed under the MIT software license, see the accompanying
-// file COPYING or http://www.opensource.org/licenses/mit-license.php.
-#include "validationinterface.h"
+// file COPYING or https://www.opensource.org/licenses/mit-license.php.
+#include <validationinterface.h>
 
 class CACNotificationInterface : public CValidationInterface
 {
