+++ src/mnode/mnode-validation.h
@@ -1,15 +1,14 @@
 #pragma once
-// Copyright (c) 2018-2021 The Pastel Core developers
+// Copyright (c) 2018-2022 The Pastel Core developers
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or https://www.opensource.org/licenses/mit-license.php.
 #include <string>
 
-#include "main.h"
+#include <main.h>
 
 #ifdef ENABLE_WALLET
-#include "wallet/wallet.h"
+#include <wallet/wallet.h>
 #endif
-
 
 bool GetBlockHash(uint256& hashRet, int nBlockHeight);
 bool GetUTXOCoin(const COutPoint& outpoint, CCoins& coins);
