+++ src/mnode/rpc/governance-rpc.h
@@ -1,0 +1,11 @@
+#pragma once
+// Copyright (c) 2018-2022 The Pastel Core developers
+// Distributed under the MIT/X11 software license, see the accompanying
+// file COPYING or https://www.opensource.org/licenses/mit-license.php.
+
+#include <univalue.h>
+#include <mnode/tickets/ticket-types.h>
+
+#ifdef GOVERNANCE_TICKETS
+UniValue governance(const UniValue& params, bool fHelp);
+#endif // GOVERNANCE_TICKETS
