+++ src/mnode/rpc/mnode-rpc.cpp
@@ -71,135 +71,6 @@ UniValue mnsync(const UniValue& params, bool fHelp)
     return "failure";
 }
 
-UniValue governance(const UniValue& params, bool fHelp)
-{
-    string strMode;
-    if (!params.empty())
-        strMode = params[0].get_str();
-
-       if (fHelp || (strMode != "ticket" && strMode != "list"))
-            throw runtime_error(
-R"(governance [ticket|list]
-
-Cast a governance vote for new or existing ticket.
-
-Examples:
-)"
-+ HelpExampleCli("governance", "")
-+ HelpExampleRpc("governance", "")
-);
-
-    string strCmd, strError;
-    if (strMode == "ticket")
-    {
-        if (params.size() < 4 || params.size() > 6)
-            throw JSONRPCError(RPC_INVALID_PARAMETER,
-R"(1.
-governance ticket add "address" amount "note" <yes|no>
-2.
-governance ticket vote "ticketID" <yes|no>
-)"
-);
-
-        UniValue resultObj(UniValue::VOBJ);
-    
-        strCmd = params[1].get_str();
-        if (strCmd == "add")
-        {
-            if (params.size() != 6)
-                throw JSONRPCError(RPC_INVALID_PARAMETER,
-R"(governance ticket add "address" amount "note" <yes|no>)");
-
-            string address = params[2].get_str();
-            CAmount amount = get_number(params[3]) * COIN;
-            string note = params[4].get_str();
-            string vote = params[5].get_str();
-
-            if (vote != "yes" && vote != "no")
-                throw JSONRPCError(RPC_INVALID_PARAMETER,
-R"(governance ticket add "address" amount "note" <yes|no>)");
-
-            uint256 newTicketId;
-            if (!masterNodeCtrl.masternodeGovernance.AddTicket(address, amount, note, (vote == "yes"), newTicketId, strError)) {
-                resultObj.pushKV(RPC_KEY_RESULT, RPC_RESULT_FAILED);
-                resultObj.pushKV(RPC_KEY_ERROR_MESSAGE, strError);
-            } else {
-                resultObj.pushKV(RPC_KEY_RESULT, RPC_RESULT_SUCCESS);
-                resultObj.pushKV("ticketId", newTicketId.ToString());
-            }
-            return resultObj;
-        }
-        if (strCmd == "vote")
-        {
-            if (params.size() != 4)
-                throw JSONRPCError(RPC_INVALID_PARAMETER,
-R"(governance ticket vote "ticketID" <yes|no>)");
-
-            string ticketIdstr = params[2].get_str();
-            string vote = params[3].get_str();
-
-            if (vote != "yes" && vote != "no")
-                throw JSONRPCError(RPC_INVALID_PARAMETER,
-R"(governance ticket add "address" amount "note" <yes|no>)");
-
-            if (!IsHex(ticketIdstr))
-                throw JSONRPCError(RPC_INVALID_PARAMETER,
-					"Invalid parameter, expected hex ticketId");
-
-            uint256 ticketId = uint256S(ticketIdstr);
-
-            if (!masterNodeCtrl.masternodeGovernance.VoteForTicket(ticketId, (vote == "yes"), strError)) {
-                resultObj.pushKV(RPC_KEY_RESULT, RPC_RESULT_FAILED);
-                resultObj.pushKV(RPC_KEY_ERROR_MESSAGE, strError);
-            } else {
-                resultObj.pushKV(RPC_KEY_RESULT, RPC_RESULT_SUCCESS);
-            }
-            return resultObj;
-        }
-    }
-
-    if(strMode == "list")
-    {
-        UniValue resultArray(UniValue::VARR);
-    
-        if (params.size() != 2)
-            throw JSONRPCError(RPC_INVALID_PARAMETER,
-R"(1.
-governance list tickets
-2.
-governance list winners)"
-);
-        strCmd = params[1].get_str();
-        if (strCmd == "tickets")
-        {
-            for (auto& s : masterNodeCtrl.masternodeGovernance.mapTickets) {
-                string id = s.first.ToString();
-
-                UniValue obj(UniValue::VOBJ);
-                obj.pushKV("id", id);
-                obj.pushKV("ticket", s.second.ToString());
-                resultArray.push_back(obj);
-            }
-        }
-        if (strCmd == "winners")
-        {
-            for (auto& s : masterNodeCtrl.masternodeGovernance.mapTickets) {
-                if (s.second.nLastPaymentBlockHeight != 0) {
-                    string id = s.first.ToString();
-    
-                    UniValue obj(UniValue::VOBJ);
-                    obj.pushKV("id", id);
-                    obj.pushKV("ticket", s.second.ToString());
-                    resultArray.push_back(obj);
-                }
-            }
-        }
-
-        return resultArray;
-    }
-    return NullUniValue;
-}
-
 UniValue getfeeschedule(const UniValue& params, bool fHelp)
 {
     if (fHelp)
@@ -380,7 +251,9 @@ static const CRPCCommand commands[] =
     { "mnode",               "masternodelist",         &masternodelist,         true  },
     { "mnode",               "masternodebroadcast",    &masternodebroadcast,    true  },
     { "mnode",               "mnsync",                 &mnsync,                 true  },
+#ifdef GOVERNANCE_TICKETS
     { "mnode",               "governance",             &governance,             true  },
+#endif // GOVERNANCE_TICKETS
     { "mnode",               "pastelid",               &pastelid,               true  },
     { "mnode",               "storagefee",             &storagefee,             true  },
     { "mnode",               "getfeeschedule",         &getfeeschedule,         true  },
