+++ src/mnode/rpc/pastelid-rpc.cpp
@@ -1,12 +1,13 @@
-// Copyright (c) 2018-2021 The Pastel Core developers
+// Copyright (c) 2018-2022 The Pastel Core developers
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or https://www.opensource.org/licenses/mit-license.php.
 
-#include "pastelid/pastel_key.h"
-#include "rpc/rpc_parser.h"
-#include "rpc/rpc_consts.h"
-#include "rpc/server.h"
-#include "mnode/rpc/pastelid-rpc.h"
+#include <pastelid/pastel_key.h>
+#include <rpc/rpc_parser.h>
+#include <rpc/rpc_consts.h>
+#include <rpc/server.h>
+#include <mnode/rpc/pastelid-rpc.h>
+#include <mnode/rpc/governance-rpc.h>
 
 using namespace std;
